Use correct path `/var/lib/kubelet` to mount kubelet data volume
