+++ pkg/component/extensions/operatingsystemconfig/executor/executor_test.go
@@ -26,7 +26,7 @@ import (
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/downloader"
 	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/executor"
-	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/original/components/varlibmount"
+	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/original/components/varlibkubeletmount"
 	"github.com/gardener/gardener/pkg/utils"
 	"github.com/gardener/gardener/pkg/utils/imagevector"
 )
@@ -58,7 +58,7 @@ var _ = Describe("Executor", func() {
 				"docker.service",
 				"unit1",
 				"unit2",
-				varlibmount.UnitName,
+				varlibkubeletmount.UnitName,
 				"unit3",
 				downloader.UnitName,
 				"unit4",
@@ -388,7 +388,7 @@ if ! diff "$PATH_CLOUDCONFIG" "$PATH_CLOUDCONFIG_OLD" >/dev/null || \
 `
 
 	for _, name := range units {
-		if name != "docker.service" && name != varlibmount.UnitName && name != downloader.UnitName {
+		if name != "docker.service" && name != varlibkubeletmount.UnitName && name != downloader.UnitName {
 			footerPart += `    systemctl enable ` + name + ` && systemctl restart --no-block ` + name + `
 `
 		}
