+++ pkg/component/extensions/operatingsystemconfig/original/original_test.go
@@ -116,7 +116,7 @@ var _ = Describe("Original", func() {
 
 			Expect(order).To(Equal([]string{
 				"valitail",
-				"var-lib-mount",
+				"var-lib-kubelet-mount",
 				"root-certificates",
 				"containerd",
 				"containerd-initializer",
@@ -137,7 +137,7 @@ var _ = Describe("Original", func() {
 
 			Expect(order).To(Equal([]string{
 				"valitail",
-				"var-lib-mount",
+				"var-lib-kubelet-mount",
 				"root-certificates",
 				"containerd",
 				"containerd-initializer",
@@ -159,7 +159,7 @@ var _ = Describe("Original", func() {
 
 			Expect(order).To(Equal([]string{
 				"valitail",
-				"var-lib-mount",
+				"var-lib-kubelet-mount",
 				"root-certificates",
 				"containerd",
 				"containerd-initializer",
